Refactor component imports, enhance loading and empty states, and introduce ServerCard component
