+++ apps/electron/src/renderer/components/TitleBar.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useWorkspaceStore } from "@/renderer/stores/workspace-store";
-import { WorkspaceSwitcher } from "./WorkspaceSwitcher";
+import { WorkspaceSwitcher } from "./workspace/WorkspaceSwitcher";
 import { usePlatformAPI } from "@/main/infrastructure/platform-api";
 
 export function TitleBar() {
