+++ apps/electron/src/renderer/components/workspace/WorkspaceSwitcher.tsx
